\name{returnALetter-package}
\alias{returnALetter-package}
\alias{returnALetter}
\docType{package}
\title{
\packageTitle{returnALetter}
}
\description{
\packageDescription{returnALetter}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{returnALetter}
\packageIndices{returnALetter}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{returnALetter}

Maintainer: \packageMaintainer{returnALetter}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
